Update libfabric_rail_manager.cpp
Temp fix to get around a segv/core.  numa_efa_devices return 12 entries all of which point to the same rail.  Deregister memory cores when it tries to deregister memory multiple times from the same rail.  We will update this fix once we have a better understanding of how to manage the numa_efa_devices to rail mapping.